[ci] Add a `redpen` check
`redpen` (https://github.com/estebank/redpen) is a Rust linter that
allows us to detect unintentional panics in our code. This PR adds a
step to the CI workflow that installs and invokes `redpen`.

Future PRs can make use of `redpen` to reduce the number of
potentially-panicing functions in `zerocopy`.